Fixed multi test. Added missing source image. Renamed files to snake_case
