+++ include/seam_carver.h
